Quote git_dir when used in git command execution
This prevents exceptions (and possibly worse) when attempting to execute git against repositories whose path contains a space.
Note that this doesn't try to protect against stranger cases, like directory names that contain double-quotes, but I'm not familiar with all the special characters that would need to be escaped here.