Reporter no longer runs SimpleCov, but aggregates/posts results instead
Due to a history of legacy functionality changes with the reporter and
on codeclimate.com, this reporter had one major flaw that because of
the circumstances didn't really bother anyone: if tests failed or errors
occurred during a test run, the reporter would still report coverage
to codeclimate.com. This has started to become a problem since we have
released a browser extension and other features which leverage test reports
on non-master or 'default' branches.

The solution is to make this reporter similar to the others for the major
languages we support. Instead of including a snippet in a test/spec helper
which initiates and runs SimpleCov and then posts results to codeclimate,
users will now do what they did before: include the SimpleCov snippet in
their helpers. After tests, users will execute the added bin/codeclimate-ruby
binary, which will aggregate and format results before posting them to
codeclimate. This obviates the "partial results" problem because your CI
script will not reach the line where you run bin/codeclimate-ruby if there
is an error or failure.

We will now have parity between how this reporter is used and how others
are used, and the presenting issues should be mitigated.