Stop sending the test framework
It's something that we can no longer infer, now that we're running our
test reporter code separately from the tests themselves.

Or, if we want to infer it, we can't infer it based on the current
process, which is what simplecov will try to do:

https://github.com/colszowka/simplecov/blob/6e5ec2342cc1cfea16ca6c3bb1034062267defbf/lib/simplecov/command_guesser.rb#L43-L56

I don't _think_ that we need to include this, and I'm trying to verify
how we use it.