Pass thru Travis PR SHA
When Travis builds a PR, as opposed to a push, it checks out the merge
commit that GitHub exposes and runs the tests at that SHA.

When we report the results, however, we'd like to report them for the
HEAD SHA from the PR, because otherwise the status won't show up.

it's possible the coverage results will differ slightly between the
merge SHA and the HEAD sha, but I'm not sure if anyone will be offended
by this fiction. We've had one report that a user is surprised not to
receive a status check when they only have Pull Request builds enabled,
and not Push builds.