Add a changelog
Also updated the release process (mostly copied from minidoc) to
emphasize the importance of updating the changelog

Moved the 1.0.0 upgrade instructions into the changelog. Left the
section, because we linked to it in some places, and I don't want to
break those links.

Didn't document the changelog going back all the way to the beginning.
We could fill that in, but I don't think it's urgent.